# ClientStubGenerator/Input/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Auction API",
    "description": "The set of supported endpoints for integrating with the amazing FOOBAR auction site®",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/food-stock/auctions/{auctionId}/items": {
      "post": {
        "tags": [
          "Auction"
        ],
        "summary": "Register an item to be sold at the given auction.",
        "operationId": "PostItemForAuction",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewItemForAuction.Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostNewItemForAuction.Response"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Auction"
        ],
        "summary": "Get all items for sale at the auction defined by auctionId.",
        "operationId": "GetAuctionedItems",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAuctionedItems.Response"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Common.RiceQuality": {
        "enum": [
          "Diamond",
          "Gold",
          "Silver",
          "Bronze",
          "Cobber",
          "Iron"
        ],
        "type": "string"
      },
      "GetAuctionedItems.Response": {
        "required": [
          "itemId",
          "quantityKg",
          "riceQuality"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid",
            "example": "f72c083d-6df7-4130-8c87-0662186eae3f"
          },
          "minimumBidDkk": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 999
          },
          "quantityKg": {
            "type": "number",
            "format": "double",
            "example": 100
          },
          "riceQuality": {
            "$ref": "#/components/schemas/Common.RiceQuality"
          }
        },
        "additionalProperties": false
      },
      "PostNewItemForAuction.Request": {
        "required": [
          "quantityKg",
          "riceQuality",
          "sellerId"
        ],
        "type": "object",
        "properties": {
          "sellerId": {
            "pattern": ".{3}.*",
            "type": "string",
            "example": "606"
          },
          "minimumBidDkk": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "example": 250
          },
          "quantityKg": {
            "type": "number",
            "format": "double",
            "example": 5
          },
          "riceQuality": {
            "$ref": "#/components/schemas/Common.RiceQuality"
          },
          "sellerName": {
            "type": "string",
            "nullable": true,
            "example": "Johnson Groceries VA"
          }
        },
        "additionalProperties": false
      },
      "PostNewItemForAuction.Response": {
        "required": [
          "itemId",
          "registrationTime",
          "sellerId"
        ],
        "type": "object",
        "properties": {
          "sellerId": {
            "type": "string"
          },
          "sellerName": {
            "type": "string",
            "nullable": true
          },
          "itemId": {
            "type": "string",
            "readOnly": true
          },
          "registrationTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}